# Conflicts:
#	src/main/java/org/ddolib/ddo/core/CompilationInput.java
#	src/main/java/org/ddolib/ddo/core/DecisionDiagram.java
#	src/main/java/org/ddolib/ddo/core/Solver.java
#	src/main/java/org/ddolib/ddo/examples/tsp/TSPAdjacentEdgeIncremental.java
#	src/main/java/org/ddolib/ddo/implem/mdd/LinkedDecisionDiagram.java
#	src/main/java/org/ddolib/ddo/implem/solver/ParallelSolver.java

+++ src/main/java/org/ddolib/ddo/core/CompilationInput.java
@@ -46,7 +46,7 @@ public final class CompilationInput<T, K> {
     /**
      * The best known lower bound at the time when the dd is being compiled
      */
-    final int bestLB;
+    final double bestLB;
 
     /**
      * The type of cut set to be used in the compilation
@@ -82,7 +82,7 @@ public final class CompilationInput<T, K> {
             final SubProblem<T> residual,
             final int maxWidth,
             final DominanceChecker<T, K> dominance,
-            final int bestLB,
+            final double bestLB,
             final CutSetType cutSetType,
             final boolean exportAsDot
     ) {
@@ -151,7 +151,7 @@ public final class CompilationInput<T, K> {
     /**
      * @return best known lower bound at the time when the dd is being compiled
      */
-    public int getBestLB() {
+    public double getBestLB() {
         return bestLB;
     }
