+++ src/main/java/org/ddolib/ddo/core/DecisionDiagram.java
@@ -25,15 +25,9 @@ public interface DecisionDiagram<T, K> {
      * @return true iff the diagram resulting from the compilation is an exact dd
      */
     boolean isExact();
-
-    /**
-     * @return the value of the best solution in this decision diagram if there is one
-     */
-    Optional<Integer> bestValue();
-
-    /**
-     * @return the solution leading to the best solution in this decision diagram (if it exists)
-     */
+    /** @return the value of the best solution in this decision diagram if there is one */
+    Optional<Double> bestValue();
+    /** @return the solution leading to the best solution in this decision diagram (if it exists) */
     Optional<Set<Decision>> bestSolution();
 
     /**
