+++ src/main/java/org/ddolib/ddo/core/Solver.java
@@ -30,7 +30,7 @@ public interface Solver {
     /**
      * @return the value of the best solution in this decision diagram if there is one
      */
-    Optional<Integer> bestValue();
+    Optional<Double> bestValue();
 
     /**
      * @return the solution leading to the best solution in this decision diagram (if it exists)
