+++ src/test/java/org/ddolib/ddo/examples/tstptw/TSPTWTests.java
@@ -52,10 +52,10 @@ public class TSPTWTests {
             vars.add(i);
         }
 
-        int rub = relax.fastUpperBound(problem.initialState(), vars);
+        double rub = relax.fastUpperBound(problem.initialState(), vars);
         // Checks if the upper bound at the root is bigger than the optimal solution
         assertTrue(rub >= instance.optimal.get(),
-                String.format("Upper bound %d is not bigger than the expected optimal solution %d",
+                String.format("Upper bound %.1f is not bigger than the expected optimal solution %.1f",
                         rub,
                         instance.optimal.get()));
     }
